+++ py_experimenter/config.py
@@ -1,7 +1,7 @@
 import logging
 from abc import ABC, abstractclassmethod
 from logging import Logger
-from typing import Any, Dict, List, Tuple, Union
+from typing import Any, Dict, List, Optional, Tuple, Union
 
 import numpy as np
 import omegaconf
@@ -9,7 +9,10 @@ from attr import dataclass
 from omegaconf import DictConfig, ListConfig, OmegaConf
 
 from py_experimenter import utils
-from py_experimenter.exceptions import InvalidColumnError, InvalidConfigError, InvalidLogtableError
+from py_experimenter.exceptions import (
+    InvalidColumnError,
+    InvalidLogtableError,
+)
 
 
 class Cfg(ABC):
@@ -41,7 +44,6 @@ class DatabaseCfg(Cfg):
     def __init__(
         self,
         provider: str,
-        use_ssh_tunnel: bool,
         database_name: str,
         table_name: str,
         result_timestamps: bool,
@@ -55,8 +57,6 @@ class DatabaseCfg(Cfg):
 
         :param provider: Database Provider; either `sqlite` or `mysql`
         :type provider: str
-        :param use_ssh_tunnel: Whether to use an SSH tunnel to connect to the database
-        :type use_ssh_tunnel: bool
         :param database_name: Name of the database
         :type database_name: str
         :param table_name: Name of the table
@@ -70,7 +70,6 @@ class DatabaseCfg(Cfg):
         :type logtables: Dict[str, Dict[str,str]]
         """
         self.provider = provider
-        self.use_ssh_tunnel = use_ssh_tunnel
         self.database_name = database_name
         self.table_name = table_name
         self.result_timestamps = result_timestamps
@@ -85,8 +84,6 @@ class DatabaseCfg(Cfg):
         database_config = config["PY_EXPERIMENTER"]["Database"]
         table_config = database_config["table"]
         provider = database_config["provider"]
-        # Optional use_ssh_tunnel
-        use_ssh_tunnel = database_config["use_ssh"] if "use_ssh" in database_config else False
         database_name = database_config["database"]
         table_name = database_config["table"]["name"]
 
@@ -99,7 +96,6 @@ class DatabaseCfg(Cfg):
 
         return DatabaseCfg(
             provider,
-            use_ssh_tunnel,
             database_name,
             table_name,
             result_timestamps,
@@ -210,9 +206,6 @@ class DatabaseCfg(Cfg):
     def valid(self) -> bool:
         if self.provider not in ["sqlite", "mysql"]:
             self.logger.error("Database provider must be either sqlite or mysql")
-            return False
-        if self.use_ssh_tunnel not in [True, False]:
-            self.logger.error("Use SSH tunnel must be a boolean.")
             return False
         if not isinstance(self.database_name, str):
             self.logger.error("Database name must be a string")
@@ -362,8 +355,10 @@ class PyExperimenterCfg:
         self.codecarbon_configuration = codecarbon_configuration
 
     @staticmethod
-    def extract_config(config_path: str, logger: logging.Logger) -> "PyExperimenterCfg":
+    def extract_config(config_path: str, logger: logging.Logger, overwritten_table_name: Optional[str] = None) -> "PyExperimenterCfg":
         config = omegaconf.OmegaConf.load(config_path)
+        if overwritten_table_name is not None:
+            config["PY_EXPERIMENTER"]["Database"]["table"]["name"] = overwritten_table_name
 
         if "n_jobs" not in config["PY_EXPERIMENTER"]:
             config["PY_EXPERIMENTER"]["n_jobs"] = 1