+++ test/test_logtables/test_mysql.py
@@ -1,11 +1,14 @@
 import logging
-import os
 from math import cos, sin
+import os
+import time
+from unittest.mock import MagicMock, patch
 
 from freezegun import freeze_time
 from mock import MagicMock, call, patch
 from omegaconf import OmegaConf
 
+from py_experimenter import database_connector_mysql
 from py_experimenter.config import DatabaseCfg
 from py_experimenter.database_connector import DatabaseConnector
 from py_experimenter.database_connector_mysql import DatabaseConnectorMYSQL
@@ -16,7 +19,6 @@ from py_experimenter.result_processor import ResultProcessor
 @patch("py_experimenter.experimenter.DatabaseConnectorMYSQL._create_database_if_not_existing")
 @patch("py_experimenter.experimenter.DatabaseConnectorMYSQL._test_connection")
 @patch("py_experimenter.experimenter.DatabaseConnectorMYSQL.fill_table")
-@patch("py_experimenter.experimenter.DatabaseConnectorMYSQL.start_ssh_tunnel")
 @patch("py_experimenter.experimenter.DatabaseConnectorMYSQL.connect")
 @patch("py_experimenter.experimenter.DatabaseConnectorMYSQL.cursor")
 @patch("py_experimenter.experimenter.DatabaseConnectorMYSQL.fetchall")
@@ -29,7 +31,6 @@ def test_tables_created(
     cursor_mock,
     connect_mock,
     fill_table_mock,
-    shh_tunnel_mock,
     test_connection_mock,
     create_database_mock,
 ):
@@ -39,7 +40,6 @@ def test_tables_created(
     cursor_mock.return_value = None
     connect_mock.return_value = None
     fill_table_mock.return_value = None
-    shh_tunnel_mock.return_value = None
     create_database_mock.return_value = None
     test_connection_mock.return_value = None
     experimenter = PyExperimenter(os.path.join("test", "test_logtables", "mysql_logtables.yml"))
@@ -57,10 +57,8 @@ def test_tables_created(
     )
 
 
-@patch("py_experimenter.database_connector_mysql.DatabaseConnectorMYSQL.start_ssh_tunnel")
 @freeze_time("2012-01-14 03:21:34")
-def test_logtable_insertion(ssh_mock):
-    ssh_mock.return_value = None
+def test_logtable_insertion():
     logger = logging.getLogger("test_logger")
     config = OmegaConf.load(os.path.join("test", "test_logtables", "mysql_logtables.yml"))
     config = DatabaseCfg.extract_config(config, logger)
@@ -74,7 +72,6 @@ def test_logtable_insertion(ssh_mock):
 
 @patch("py_experimenter.experimenter.DatabaseConnectorMYSQL._create_database_if_not_existing")
 @patch("py_experimenter.experimenter.DatabaseConnectorMYSQL._test_connection")
-@patch("py_experimenter.experimenter.DatabaseConnectorMYSQL.start_ssh_tunnel")
 @patch("py_experimenter.experimenter.DatabaseConnectorMYSQL.connect")
 @patch("py_experimenter.experimenter.DatabaseConnectorMYSQL.cursor")
 @patch("py_experimenter.experimenter.DatabaseConnectorMYSQL.commit")
@@ -88,15 +85,12 @@ def test_delete_logtable(
     fetchall_mock,
     cursor_mock,
     connect_mock,
-    ssh_mock,
     test_connection_mock,
     create_database_mock,
 ):
     fetchall_mock.return_value = cursor_mock.return_value = connect_mock.return_value = commit_mocck.return_value = None
-    close_connection_mock.return_value = test_connection_mock.return_value = create_database_mock.return_value = execution_mock.return_value = (
-        ssh_mock.return_value
-    ) = None
-    experimenter = PyExperimenter(os.path.join("test", "test_logtables", "mysql_logtables.yml"), use_codecarbon=False, use_ssh_tunnel=False)
+    close_connection_mock.return_value = test_connection_mock.return_value = create_database_mock.return_value = execution_mock.return_value = None
+    experimenter = PyExperimenter(os.path.join("test", "test_logtables", "mysql_logtables.yml"), use_codecarbon=False)
     experimenter.delete_table()
     execution_mock.assert_has_calls(
         [
@@ -143,7 +137,6 @@ def test_integration_with_resultfields():
     logtable2 = [x[:2] + x[3:] for x in logtable2]
     assert logtable2 == [(1, 1, 1), (2, 1, 3)]
     assert timesteps == timesteps2
-    experimenter.close_ssh()
 
 
 # Integration Test without Resultfields
@@ -156,7 +149,7 @@ def own_function_without_resultfields(keyfields: dict, result_processor: ResultP
 
 
 def test_integration_without_resultfields():
-    experimenter = PyExperimenter(os.path.join("test", "test_logtables", "mysql_logtables_no_resultfields.yml"), use_ssh_tunnel=False)
+    experimenter = PyExperimenter(os.path.join("test", "test_logtables", "mysql_logtables_no_resultfields.yml"))
     try:
         experimenter.delete_table()
     except Exception:
@@ -176,4 +169,27 @@ def test_integration_without_resultfields():
     logtable2 = [x[:2] + x[3:] for x in logtable2]
     assert logtable2 == [(1, 1, 1), (2, 1, 3)]
     assert timesteps == timesteps2
-    experimenter.close_ssh()
+
+def own_function_without_resultfields_stagger(keyfields: dict, result_processor: ResultProcessor, custom_fields: dict):
+    result_processor.process_logs({"log": {"test": 0}})
+    result_processor.process_logs({"log": {"test": 2}})
+    time.sleep(10)
+    result_processor.process_logs({"log": {"test": 4}})
+
+
+def test_stagger_logging():
+    experimenter = PyExperimenter(
+        os.path.join("test", "test_logtables", "mysql_logtables.yml"),
+        stagger_logging=True,
+        log_every_n_seconds=10,
+    )
+    try:
+        experimenter.delete_table()
+    except Exception:
+        pass
+
+    experimenter.fill_table_from_config()
+    experimenter.execute(own_function_without_resultfields_stagger, max_experiments=1)
+
+    logtable = experimenter.get_logtable("log")
+    assert (logtable.groupby("timestamp").count()["test"] == [2,1]).all()