+++ CHANGELOG.rst
@@ -1,7 +1,19 @@
 =========
 Changelog
 =========
+v1.4.3 (**.**.2025)
+===================
 
+Feature
+-------
+
+- Added stagger_logging functionality to logtables.
+- Allow condition for experimenter.get_table() and experimenter.get_logtable()
+
+Fix
+===
+
+- Fix bug, where the logtable_name was not overwritten by `table_name` updates in the `PyExperimenter` class.
 
 v1.4.2 (12.06.2024)
 ===================